+++ lib/sensors/yahabme280.cpp
@@ -8,6 +8,8 @@
  * @brief
  * Provides a class to connect to WLAN
  */
+
+#define __DEBUG
 #include "debug.h"
 #include "message.h"
 #include "yahabme280.h"