+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-# change log for yaha weather station
+# change log for yaha station
+
+## 0.2.0 2020-05-31 update
+
+- Added functionality for irrigation
+- Added switch to enable/disable battery mode
 
 ## 0.1.0 2020-05-31 Initial revision